toolkit/components/passwordmgr/test/mochitest/test_xml_load.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test XML document prompts</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <script type="text/javascript" src="../../../prompts/test/prompt_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: XML prompt
<p id="display"></p>

<div id="content" style="display: none">
  <iframe id="iframe"></iframe>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Login Manager: XML prompts. **/
// Force parent to not look for tab-modal prompts, as they're not used for auth prompts.
// Used by prompt_common.js.
isTabModal = false;

function makeRequest(uri, expectedUser, expectedPass) {
  let xmlDoc = document.implementation.createDocument("", "test", null);
  xmlDoc.load(uri);
  return new Promise(resolve => {
    xmlDoc.addEventListener("load", documentLoaded);

    function documentLoaded(e) {
      let username = xmlDoc.getElementById("user").textContent;
      let password = xmlDoc.getElementById("pass").textContent;
      is(username, expectedUser, "Checking provided username");
      is(password, expectedPass, "Checking provided password");
      resolve();
    }
  });
}

function checkWindowFocus() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  // Check that the right tab is focused:
  let browserWin = Services.wm.getMostRecentWindow("navigator:browser");
  let spec = browserWin.gBrowser.selectedBrowser.currentURI.spec;
  sendAsyncMessage("focusResult", spec.startsWith("http://mochi.test:8888"));
}

runInParent(function initLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  let login1 = Cc["@mozilla.org/login-manager/loginInfo;1"]
                        .createInstance(Ci.nsILoginInfo);
  let login2 = Cc["@mozilla.org/login-manager/loginInfo;1"]
                        .createInstance(Ci.nsILoginInfo);

  login1.init("http://mochi.test:8888", null, "xml",
              "xmluser1", "xmlpass1", "", "");
  login2.init("http://mochi.test:8888", null, "xml2",
              "xmluser2", "xmlpass2", "", "");

  Services.logins.addLogin(login1);
  Services.logins.addLogin(login2);
});

add_task(async function prepare() {
  await new Promise(resolve => SpecialPowers.pushPrefEnv({"set": [
    ["dom.xmldocument.load.enabled", true],
  ]}, resolve));
});

add_task(async function test_promptUsernameAndPassword_accept() {
  let loadPromise = makeRequest("authenticate.sjs?user=xmluser1&pass=xmlpass1&realm=xml", "xmluser1", "xmlpass1");
  let state = {
    msg: `http://mochi.test:8888 is requesting your username and password. The site says: “xml”`,
    title: "Authentication Required",
    textValue: "xmluser1",
    passValue: "xmlpass1",
    iconClass: "authentication-icon question-icon",
    titleHidden: true,
    textHidden: false,
    passHidden: false,
    checkHidden: true,
    checkMsg: "",
    checked: false,
    focused: "textField",
    defButton: "button0",
  };

  let action = {
    buttonClick: "ok",
  };

  await handlePrompt(state, action);
  await loadPromise;
});

add_task(async function test_backgroundTab() {
  // Test correct parenting, by opening another tab in the foreground,
  // and making sure the prompt re-focuses the original tab when shown:
  let newWin = window.open();
  newWin.focus();
  let loadPromise = makeRequest("authenticate.sjs?user=xmluser2&pass=xmlpass2&realm=xml2", "xmluser2", "xmlpass2");

  // We do this because the handlePrompt function ensures that the dialog exists before proceeding with other checks.
  let state = {
    msg: `http://mochi.test:8888 is requesting your username and password. The site says: “xml2”`,
    title: "Authentication Required",
    textValue: "xmluser2",
    passValue: "xmlpass2",
    iconClass: "authentication-icon question-icon",
    titleHidden: true,
    textHidden: false,
    passHidden: false,
    checkHidden: true,
    checkMsg: "",
    checked: false,
    focused: "textField",
    defButton: "button0",
    chrome: true,
    dialog: true,
    chromeDependent: true,
    isWindowModal: true,
  };

  let action = {
    buttonClick: "none",
  };

  await handlePrompt(state, action);

  await new Promise(resolve => {
    let focusScript = runInParent(checkWindowFocus);
    focusScript.addMessageListener("focusResult", (msg) => {
      focusScript.removeMessageListener("focusResult");
      ok(msg, "Dialog appears on the right tab");
      resolve(true);
    });
  });

  action = {
    buttonClick: "ok",
  };

  await handlePrompt(state, action);
  await loadPromise;
  newWin.close();
});
</script>
</pre>
</body>
</html>
